Revert "Gutted project so I can start from scratch"
This reverts commit 9db4049fcde8a68ba9bcb957b65af3dcbc3c2daa.
