+++ Fuel calculator/Controls/NumericInput.xaml
@@ -7,15 +7,6 @@
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
     <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition />
-            <RowDefinition />
-        </Grid.RowDefinitions>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition />
-            <ColumnDefinition />
-        </Grid.ColumnDefinitions>
-        <Label Grid.Column="0" />
-        <TextBox Grid.Row="0" Grid.Column="1"></TextBox>
+            
     </Grid>
 </UserControl>
