+++ Fuel calculator/Logger.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Globalization;
 using System.IO;
+using System.Xml;
 
 namespace Fuel_calculator
 {
