--- Fuel calculator/MainWindow - Copy.xaml
@@ -1,71 +1,0 @@
-﻿<Window x:Class="Fuel_calculator.MainWindowc"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        mc:Ignorable="d"
-        Title="Fuel Calculator" Height="450" Width="760" SizeToContent="Manual"
-        WindowStyle="ThreeDBorderWindow" Icon="Logo.png" Background="LightGray">
-    <Viewbox Stretch="Uniform">
-        <Grid Name="MainGrid" >
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="Auto"></ColumnDefinition>
-                <ColumnDefinition Width="Auto" MinWidth="30"></ColumnDefinition>
-                <ColumnDefinition Width="Auto"></ColumnDefinition>
-                <ColumnDefinition Width="Auto" MinWidth="30"></ColumnDefinition>
-                <ColumnDefinition Width="100"></ColumnDefinition>
-            </Grid.ColumnDefinitions>
-
-            <Grid.RowDefinitions>
-                <RowDefinition Height="Auto"/>
-                <RowDefinition Height="Auto"/>
-                <RowDefinition Height="Auto"/>
-                <RowDefinition Height="Auto"/>
-                <RowDefinition Height="Auto"/>
-                <RowDefinition Height="Auto"/>
-                <RowDefinition Height="Auto"/>
-                <RowDefinition Height="Auto"/>
-            </Grid.RowDefinitions>
-
-            <Label Grid.Column="0" Grid.Row="0" Content="Track" FontWeight="Bold"></Label>
-            <ComboBox Grid.Column="1" Grid.ColumnSpan="4" Grid.Row="0" Name="TrackSelector" SelectionChanged="TrackSelector_SelectionChanged"/>
-
-
-            <Label Grid.Column="0" Grid.Row="1" Content="Total Race Time" />
-            <TextBox Grid.Column="1" Grid.Row="1" Name="TotalRaceTimeHours" TextChanged="TotalRaceTime_TextChanged"></TextBox>
-            <Label Grid.Column="2" Grid.Row="1">H</Label>
-            <TextBox Grid.Column="3" Grid.Row="1" Name="TotalRaceTimeMinutes" TextChanged="TotalRaceTime_TextChanged"></TextBox>
-            <Label Grid.Column="4" Grid.Row="1">M</Label>
-
-
-            <Label Grid.Column="0" Grid.Row="2" Content="Average lap time" />
-            <TextBox Grid.Column="1" Grid.Row="2" Name="AverageLapTimeMinutes" TextChanged="AverageLapTime_TextChanged"></TextBox>
-            <Label Grid.Column="2" Grid.Row="2">M</Label>
-            <TextBox Grid.Column="3" Grid.Row="2" Name="AverageLapTimeSeconds" TextChanged="AverageLapTime_TextChanged"></TextBox>
-            <Label Grid.Column="4" Grid.Row="2">S</Label>
-
-            <Label Grid.Column="0" Grid.Row="3" Content="Fuel Per Lap" />
-            <TextBox Grid.Column="1" Grid.Row="3" Name="FuelPerLap" TextChanged="FuelPerLap_TextChanged" LostFocus="LogFuelPerLap"></TextBox>
-            <Label Grid.Column="2" Grid.Row="3">Liters</Label>
-            <Slider Grid.Column="3" Grid.Row="3" Name="FuelPerLapSlider" Minimum="0" Maximum="6" TickFrequency="0.01" TickPlacement="BottomRight" IsSnapToTickEnabled="True"  HorizontalAlignment="Left" Margin="10.4,5.2,0,0"  VerticalAlignment="Top" Height="23" Grid.RowSpan="2" Grid.ColumnSpan="2" Width="115" ValueChanged="FuelPerLapSlider_ValueChanged" LostFocus="LogFuelPerLap"/>
-
-            <Label Grid.Column="0" Grid.Row="4" Content="Fuel Tank Capacity" />
-            <TextBox Grid.Column="1" Grid.Row="4" Name="FuelTankCapacity" TextChanged="FuelTankCapacityChanged"></TextBox>
-            <ComboBox Grid.Column="2" Grid.ColumnSpan="3" Grid.Row="4" Name="CarSelector"
-                      SelectionChanged="CarSelector_OnSelectionChanged" HorizontalAlignment="Right" Width="175" />
-
-            <Label Grid.Column="0" Grid.Row="5" FontWeight="Bold">Total laps:</Label>
-            <Label Grid.Column="1" Grid.Row="5" FontWeight="Bold" Name="TotalLaps">0</Label>
-            <Label Grid.Column="2" Grid.Row="5" FontWeight="Bold">Laps</Label>
-            <Button Grid.Column="4" Grid.Row="5" Click="OnSaveClick">Save Fuel Setup</Button>
-
-            <Label Grid.Column="0" Grid.Row="6" FontWeight="Bold" Content="Max Stint Length"/>
-            <Label Grid.Column="1" Grid.Row="6" FontWeight="Bold" Name="MaxStintLength">0</Label>
-            <Label Grid.Column="2" Grid.Row="6" FontWeight="Bold">Laps</Label>
-
-            <Label Grid.Column="0" Grid.Row="7" FontWeight="Bold">Total Fuel Needed:</Label>
-            <Label Grid.Column="1" Grid.Row="7" Name="TotalFeulNeeded" FontWeight="Bold">0</Label>
-            <Label Grid.Column="2" Grid.Row="7" FontWeight="Bold">Liters</Label>
-        </Grid>
-    </Viewbox>
-</Window>
