+++ Fuel calculator/Xml/Objects/TrackCarCombo.cs
@@ -1,4 +1,5 @@
-﻿using System.Xml.Serialization;
+﻿using System;
+using System.Xml.Serialization;
 
 namespace Fuel_calculator.Xml.Objects
 {
