+++ Xml deserializer/Xml deserializer/Xml Objects/Catalog.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Diagnostics.CodeAnalysis;
 using System.Xml.Serialization;
 
 namespace Xml_deserializer.Xml_Objects
